+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to better preserve the files' executable permissions.
     where the placeholder will map to the shortcut's dynamic app ID.
   * Paths may now use the `<winLocalAppDataLow>` placeholder.
   * GUI: Custom games can now be expanded and collapsed.
+  * GUI: Custom games now have an icon to indicate when they override a manifest entry.
 * Fixed:
   * Files on Windows network shares were not backed up correctly.
     For example, a file identified as `\\localhost\share\test.txt`